Se implementa condicional terciario con variables de entorno y endpoint quemados con el objetivo de que tome correctamente las rutas que procesan las solicitudes realizadas desde el sdk
