+++ epayco/src/main/AndroidManifest.xml
@@ -1,6 +1,5 @@
-<manifest
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    package="co.epayco.android"
-    android:versionCode="1"
-    android:versionName="1.0.0">
+<?xml version="1.0" encoding="utf-8"?>
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+>
+
 </manifest>
