--- epayco/src/main/java/co/epayco/android/models/Card.java
@@ -1,80 +1,0 @@
-package co.epayco.android.models;
-
-public class Card {
-
-    String number;
-    String month;
-    String year;
-    String cvc;
-    Boolean hasCvv;
-
-    public Card() {}
-
-    public static final String AMERICAN_EXPRESS = "American Express";
-    public static final String DISCOVER = "Discover";
-    public static final String JCB = "JCB";
-    public static final String DINERS_CLUB = "Diners Club";
-    public static final String VISA = "Visa";
-    public static final String MASTERCARD = "MasterCard";
-    public static final String UNKNOWN = "Unknown";
-
-    public static final String[] PREFIXES_AMERICAN_EXPRESS = {"34", "37"};
-    public static final String[] PREFIXES_DISCOVER = {"60", "62", "64", "65"};
-    public static final String[] PREFIXES_JCB = {"35"};
-    public static final String[] PREFIXES_DINERS_CLUB = {"300", "301", "302", "303", "304", "305", "309", "36", "38", "39"};
-    public static final String[] PREFIXES_VISA = {"4"};
-    public static final String[] PREFIXES_MASTERCARD = {
-            "2221", "2222", "2223", "2224", "2225", "2226", "2227", "2228", "2229",
-            "223", "224", "225", "226", "227", "228", "229",
-            "23", "24", "25", "26",
-            "270", "271", "2720",
-            "50", "51", "52", "53", "54", "55"
-    };
-
-    public Card(String number, String month, String year, String cvc) {
-        this.number = number;
-        this.month = month;
-        this.year = year;
-        this.cvc = cvc;
-    }
-
-    public String getNumber() {
-        return number;
-    }
-
-    public void setNumber(String number) {
-        this.number = number;
-    }
-
-    public String getMonth() {
-        return month;
-    }
-
-    public void setMonth(String month) {
-        this.month = month;
-    }
-
-    public String getYear() {
-        return year;
-    }
-
-    public void setYear(String year) {
-        this.year = year;
-    }
-
-    public String getCvc() {
-        return cvc;
-    }
-
-    public void setCvc(String cvc) {
-        this.cvc = cvc;
-    }
-
-    public Boolean getHasCvv() {
-        return hasCvv;
-    }
-
-    public void setHasCvv(Boolean hasCvv) {
-        this.hasCvv = hasCvv;
-    }
-}
