--- epayco/src/main/java/co/epayco/android/models/Plan.java
@@ -1,79 +1,0 @@
-package co.epayco.android.models;
-
-public class Plan {
-
-    String idPlan;
-    String name;
-    String description;
-    String amount;
-    String currency;
-    String interval;
-    String intervalCount;
-    String trialDays;
-
-    public Plan() {}
-
-    public String getIdPlan() {
-        return idPlan;
-    }
-
-    public void setIdPlan(String idPlan) {
-        this.idPlan = idPlan;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public String getAmount() {
-        return amount;
-    }
-
-    public void setAmount(String amount) {
-        this.amount = amount;
-    }
-
-    public String getCurrency() {
-        return currency;
-    }
-
-    public void setCurrency(String currency) {
-        this.currency = currency;
-    }
-
-    public String getInterval() {
-        return interval;
-    }
-
-    public void setInterval(String interval) {
-        this.interval = interval;
-    }
-
-    public String getIntervalCount() {
-        return intervalCount;
-    }
-
-    public void setIntervalCount(String intervalCount) {
-        this.intervalCount = intervalCount;
-    }
-
-    public String getTrialDays() {
-        return trialDays;
-    }
-
-    public void setTrialDays(String trialDays) {
-        this.trialDays = trialDays;
-    }
-}
