--- epayco/src/main/java/co/epayco/android/time/FrozenClock.java
@@ -1,14 +1,0 @@
-package co.epayco.android.time;
-
-import java.util.Calendar;
-
-public class FrozenClock extends Clock {
-
-    public static void freeze(Calendar freeze) {
-        getInstance().calendarInstance = freeze;
-    }
-
-    public static void unfreeze() {
-        getInstance().calendarInstance = null;
-    }
-}
