--- epayco/src/main/java/co/epayco/android/util/EpaycoCallback.java
@@ -1,12 +1,0 @@
-package co.epayco.android.util;
-
-
-import org.json.JSONException;
-import org.json.JSONObject;
-
-public interface EpaycoCallback {
-
-    void onSuccess(JSONObject data) throws JSONException;
-
-    void onError(Exception error);
-}
