+++ epayco/src/main/java/com/epayco/android/models/Subscription.java
@@ -1,0 +1,70 @@
+package com.epayco.android.models;
+
+public class Subscription {
+
+    String idPlan;
+    String customer;
+    String tokenCard;
+    String ip;
+    String doc_type;
+    String doc_number;
+      //Optional
+    String urlConfirmation;
+
+    public Subscription() {}
+
+    public String getIdPlan() {
+        return idPlan;
+    }
+
+    public void setIdPlan(String idPlan) {
+        this.idPlan = idPlan;
+    }
+
+    public String getCustomer() {
+        return customer;
+    }
+
+    public void setCustomer(String customer) {
+        this.customer = customer;
+    }
+
+    public String getTokenCard() {
+        return tokenCard;
+    }
+
+    public void setTokenCard(String tokenCard) {
+        this.tokenCard = tokenCard;
+    }
+
+    public String getUrlConfirmation() {
+        return urlConfirmation;
+    }
+
+    public void setUrlConfirmation(String urlConfirmation) {
+        this.urlConfirmation = urlConfirmation;
+    }
+    public String getIp() {
+        return ip;
+    }
+
+    public void setIp(String ip) {
+        this.ip = ip;
+    }
+
+       public String getDoc_type() {
+        return doc_type;
+    }
+
+    public void setDoc_type(String doc_type) {
+        this.doc_type = doc_type;
+    }
+
+    public String getDoc_number() {
+        return doc_number;
+    }
+
+    public void setDoc_number(String doc_number) {
+        this.doc_number = doc_number;
+    }
+}
