+++ epayco/src/main/java/com/epayco/android/time/FrozenClock.java
@@ -1,0 +1,14 @@
+package com.epayco.android.time;
+
+import java.util.Calendar;
+
+public class FrozenClock extends Clock {
+
+    public static void freeze(Calendar freeze) {
+        getInstance().calendarInstance = freeze;
+    }
+
+    public static void unfreeze() {
+        getInstance().calendarInstance = null;
+    }
+}
