+++ epayco/src/main/java/com/epayco/android/util/EpaycoCallback.java
@@ -1,0 +1,12 @@
+package com.epayco.android.util;
+
+
+import org.json.JSONException;
+import org.json.JSONObject;
+
+public interface EpaycoCallback {
+
+    void onSuccess(JSONObject data) throws JSONException;
+
+    void onError(Exception error);
+}
